requirements.txt
Django==5.0.4
django-allauth==0.57.0
Django-oauth-toolkit==2.3.0
Django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
oauthlib==3.2.2
gunicorn==21.2
Django-cors-headers==4.1.0
psycopg2-binary==2.9.1
requests
python-dotenv
groq
google-cloud-aiplatform
google-auth
pyotp